separating deleteBooking and removeDogsFromBooking #255

+++ staff/carlos-fernandez/project/api/logic/deleteBooking.js
@@ -3,38 +3,20 @@ import { Errors, Validator } from "common";
 
 const { Booking } = models;
 
-export default (userId, bookingId, dogId) => {
-  Validator.id(userId);
+export default ({ bookingId, userId }) => {
+  console.log("LOGICA BOOKING ID ------", bookingId);
+  console.log("LOGICA USER ID ---------", userId);
   Validator.id(bookingId);
-  Validator.id(dogId); //Validamos que dogId es un Id valido
+  Validator.id(userId);
 
-  return Booking.findById(bookingId)
-    .then((booking) => {
-      if (!booking) {
-        throw new Errors.ExistenceError("Booking not found");
-      }
-
-      if (booking.owner.toString() !== userId) {
-        throw new Errors.ExistenceError(
-          "Booking associated to this user not found"
+  return Booking.findOneAndDelete({ _id: bookingId, owner: userId }).then(
+    (deletedBooking) => {
+      if (!deletedBooking) {
+        throw new Errors.NotFoundError(
+          "Booking or booking associated to this user not found"
         );
       }
-
-      if (!booking.dogs.some((dog) => dog.toString() === dogId)) {
-        throw new Errors.ExistenceError("Dog not found in booking");
-      }
-
-      const updatedDogs = booking.dogs.filter(
-        (dog) => dog.toString() !== dogId
-      );
-
-      return Booking.findOneAndDelete(
-        { _id: bookingId },
-        { dogs: updatedDogs },
-        { new: true }
-      );
-    })
-    .catch((error) => {
-      throw error;
-    });
+      return deletedBooking;
+    }
+  );
 };
